fix: address CodeRabbit findings in native libs
- json.c: guard escape/unicode reads against OOB past NUL terminator
- webui_little.c: fix use-after-free/double-free in stringify_response,
  eager critical-section init, rebind replaces stale callback, OOB
  guard in payload JSON parser
- little: widen lt_table_next cursor to uint64_t (60-bit index, no wrap
  past 16M entries per bucket)
- build.ps1: don't pass -rdynamic/-ldl on macOS (no libdl)
- run-e2e.ps1: use .dylib suffix on macOS for native test libs
- api.md: update lt_table_next signature

+++ src/little.h
@@ -612,7 +612,7 @@ uint8_t lt_equals(lt_Value a, lt_Value b);
 lt_Value lt_make_table(lt_VM* vm);
 lt_Value lt_table_set(lt_VM* vm, lt_Value table, lt_Value key, lt_Value val);
 lt_Value lt_table_get(lt_VM* vm, lt_Value table, lt_Value key);
-uint8_t  lt_table_next(lt_VM* vm, lt_Value table, uint32_t* cursor, lt_Value* key, lt_Value* val);
+uint8_t  lt_table_next(lt_VM* vm, lt_Value table, uint64_t* cursor, lt_Value* key, lt_Value* val);
 uint8_t  lt_table_pop(lt_VM* vm, lt_Value table, lt_Value key);
 
 lt_Value  lt_make_array(lt_VM* vm);
@@ -654,7 +654,7 @@ struct lt_Api {
 
 	lt_Value (*table_set)(lt_VM* vm, lt_Value table, lt_Value key, lt_Value val);
 	lt_Value (*table_get)(lt_VM* vm, lt_Value table, lt_Value key);
-	uint8_t (*table_next)(lt_VM* vm, lt_Value table, uint32_t* cursor, lt_Value* key, lt_Value* val);
+	uint8_t (*table_next)(lt_VM* vm, lt_Value table, uint64_t* cursor, lt_Value* key, lt_Value* val);
 	uint8_t (*table_pop)(lt_VM* vm, lt_Value table, lt_Value key);
 
 	lt_Value (*array_push)(lt_VM* vm, lt_Value array, lt_Value val);
